+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Campaign Changelog
 
+## Unreleased
+### Added
+- Added reCAPTCHA spam protection to mailing list subscription forms
+
 ## 1.0.0 - 2018-05-02
 - Stable release